Add util for validation keywords
